Add new `Workit/RestrictOnSend` cop
This cop will make sure that RESTRICT_ON_SEND is defined if on_send or after_send is defined.

```ruby
# bad
class FooCop
  def on_send(node)
    # ...
  end
end
# good
class FooCop
  RESTRICT_ON_SEND = %i[bad_method].freeze
  def on_send(node)
    # ...
  end
end
```

This comment rubocop/rubocop-rspec#1265 (comment)
is also mentioned in and leads to optimize the cop by RESTRICT_ON_SEND whenever possible.

However, this PR does not resolve existing cop.

If we included that much in this PR, the scale would be enormous,
and we thought it would be better to look at how to optimize each cop in detail.