Add new `Workit/NoopRescue` cop
This PR is add new `Lint/NoopRescue` cop.
This cop check for suppress or ignore checked exception.

Suppressing or ignoring exceptions hides the problem, makes it invisible, and makes it behave as if it were not a problem.
The test code becomes less reliable and the possibility of problems increases.